stage 4 - order students per course

+++ src/tracker/data/student/CourseScore.java
@@ -20,7 +20,7 @@ public class CourseScore {
         return id;
     }
 
-    public int getTotalPoints() {
+    public long getTotalPoints() {
         return points.stream().mapToInt(Integer::intValue).sum();
     }
